quantization/working/awq_quantize_py311/pyproject.toml
[project]
name = "awq-quantize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autoawq>=0.2.9",
    "torch==2.6.0",
    "transformers==4.51.3",
]
